Change in simd forward for convolutional and linear, makefile, backward in simd for convolutional
The simd forwards for convolutional and linear exhibited a small error: The loops over the parallelized index checked for j < N, not for j+4 < N, meaning vectors could (potentially) go out of bounds from the tensor. The results seemed to have been correct before but this is nonetheless unintended and potentially dangerous behaviour; it is fixed now.

The makefile now has a recipe for compiling the test.cc file.

Added a simd implementation to the convolution and linear backward pass. The convolution backward pass causes little inaccuracies in the loss values... gotta fix that later but it's too late in the evening / early in the morning now