Investigating the inaccuracy in the backward pass
It occurs when the calculation of gx in the convolutional backward pass uses simd instructions (I tested the data types float32x2_t and float32x4_t). I have no idea why since I use the same instructions I use in other parts of the code... The corresponding part is not in use and the arm simd implementation is exact. There is still a speed up present for the convolutional backward pass so it isn't that much of a problem.